Add LastUpdated property to NuGetToolModel
Introduced a new nullable `DateTime` property named `LastUpdated` in the `NuGetToolModel` class. This property tracks the last update time of the model. If `LastUpdated` is `null`, it indicates that the model has never been updated. This change helps optimize the process by preventing execution when there are no file changes upon rerunning the program.